+++ src/app/egresados/page.js
@@ -2,7 +2,7 @@
 import {
   Heading,
 } from '@chakra-ui/react';
-import FilterMenu from '../egresados/components/FormHandler'
+import FilterMenu from './components/FilterMenu'
 
 export default function Egresados() {
   return (
